Urls accessed differently.
The urls are now randomized in order, and a delay has been implemented in-between each call as to reduce the likelihood of being seen as malicious.
The domains that have been visited are also logged into an array, so that later down the line further measures may be taken to safeguard the program as a responsible web crawler.